+++ prince/mca.py
@@ -8,6 +8,7 @@ import scipy.sparse as sp
 import sklearn.base
 import sklearn.preprocessing
 import sklearn.utils
+from typing_extensions import override
 
 from prince import utils
 
@@ -97,8 +98,9 @@ class MCA(ca.CA, sklearn.base.TransformerMixin):
             X = X.reindex(columns=one_hot_columns_.union(X.columns), fill_value=False)
         return X
 
+    @utils.check_is_fitted
     def get_feature_names_out(self, input_features=None):
-        return np.arange(self.n_components_)
+        return np.arange(len(self.svd_.s))
 
     def _subset_greenacre_quantities(self):
         """Adjusted eigenvalues and total inertia for subset MCA with Greenacre correction.
@@ -135,6 +137,7 @@ class MCA(ca.CA, sklearn.base.TransformerMixin):
         return lambda_adj, lambda_t
 
     @property
+    @override
     def eigenvalues_(self):
         """Returns the eigenvalues associated with each principal component."""
         eigenvalues = super().eigenvalues_
@@ -151,6 +154,7 @@ class MCA(ca.CA, sklearn.base.TransformerMixin):
 
     @property
     @utils.check_is_fitted
+    @override
     def percentage_of_variance_(self):
         """Returns the percentage of explained inertia per principal component."""
         # Greenacre correction on a subset MCA: closed-form Benzécri assumes uniform row
@@ -174,6 +178,7 @@ class MCA(ca.CA, sklearn.base.TransformerMixin):
         return super().percentage_of_variance_
 
     @utils.check_is_dataframe_input
+    @override
     def fit(self, X, y=None):
         """Fit the MCA on a categorical dataframe.
 
@@ -262,12 +267,14 @@ class MCA(ca.CA, sklearn.base.TransformerMixin):
 
     @utils.check_is_dataframe_input
     @utils.check_is_fitted
+    @override
     def row_coordinates(self, X):
         """Row principal coordinates in the MCA space."""
         return super().row_coordinates(self._prepare(X))
 
     @utils.check_is_dataframe_input
     @utils.check_is_fitted
+    @override
     def row_cosine_similarities(self, X):
         """Squared cosine similarities (cos2) of each row on each component."""
         oh = self._prepare(X)
@@ -275,12 +282,14 @@ class MCA(ca.CA, sklearn.base.TransformerMixin):
 
     @utils.check_is_dataframe_input
     @utils.check_is_fitted
+    @override
     def column_coordinates(self, X):
         """Column (category) principal coordinates in the MCA space."""
         return super().column_coordinates(self._prepare(X))
 
     @utils.check_is_dataframe_input
     @utils.check_is_fitted
+    @override
     def column_cosine_similarities(self, X):
         """Squared cosine similarities (cos2) of each column on each component."""
         oh = self._prepare(X)